+++ app/lib/package/models.dart
@@ -202,12 +202,13 @@ class Package extends db.ExpandoModel<String> {
 
   // Convenience Fields:
 
-  bool get isVisible => !isBlocked;
+  bool get isVisible => !isBlocked && !(isModerated ?? false);
   bool get isNotVisible => !isVisible;
 
   bool get isIncludedInRobots {
     final now = clock.now();
     return isVisible &&
+        !(isModerated ?? false) &&
         !isDiscontinued &&
         !isUnlisted &&
         now.difference(created!) > robotsVisibilityMinAge &&
@@ -411,6 +412,14 @@ class Package extends db.ExpandoModel<String> {
     this.isBlocked = isBlocked;
     blockedReason = reason;
     blocked = isBlocked ? clock.now().toUtc() : null;
+    updated = clock.now().toUtc();
+  }
+
+  void updateIsModerated({
+    required bool isModerated,
+  }) {
+    this.isModerated = isModerated;
+    moderatedAt = isModerated ? clock.now().toUtc() : null;
     updated = clock.now().toUtc();
   }
 }